refactor: Rename package name remoteoffsetmanager to the offsetmanager
- Remove the offset-manager-local so offsetmanager name is appropriate.
