--- offset-manager-remote/src/main/java/localoffsetmanager/RemoteOffsetManagerApplication.java
@@ -1,8 +1,0 @@
-package localoffsetmanager;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class RemoteOffsetManagerApplication {
-
-}
