--- offset-manager-remote/src/main/java/localoffsetmanager/controller/dto/LastOffsetRecordBatchRequest.java
@@ -1,11 +1,0 @@
-package localoffsetmanager.controller.dto;
-
-import jakarta.validation.constraints.NotEmpty;
-
-import java.util.List;
-
-public record LastOffsetRecordBatchRequest(
-  @NotEmpty
-  List<String> keys
-) {
-}
