--- offset-manager-remote/src/main/java/localoffsetmanager/service/dto/LastOffsetRecordBatchResponse.java
@@ -1,18 +1,0 @@
-package localoffsetmanager.service.dto;
-
-import offsetmanager.domain.OffsetRecord;
-
-import java.util.List;
-
-public record LastOffsetRecordBatchResponse(
-  List<LastOffsetRecordResponse> lastOffsetRecords
-) {
-
-  public static LastOffsetRecordBatchResponse from(List<OffsetRecord> offsetRecords) {
-    return new LastOffsetRecordBatchResponse(
-      offsetRecords.stream()
-        .map(LastOffsetRecordResponse::from)
-        .toList()
-    );
-  }
-}
