--- offset-manager-remote/src/main/java/localoffsetmanager/service/dto/LastOffsetRecordResponse.java
@@ -1,15 +1,0 @@
-package localoffsetmanager.service.dto;
-
-import offsetmanager.domain.OffsetRecord;
-
-public record LastOffsetRecordResponse(
-  String key,
-  long offset
-) {
-  public static LastOffsetRecordResponse from(OffsetRecord offsetRecord) {
-    return new LastOffsetRecordResponse(
-      offsetRecord.key(),
-      offsetRecord.offset()
-    );
-  }
-}
