+++ offset-manager/build.gradle
@@ -1,0 +1,36 @@
+plugins {
+  id 'java'
+  id("org.springframework.boot") version "3.5.6"
+  id("io.spring.dependency-management") version "1.1.7"
+}
+
+group = 'sourceconnect'
+version = '1.0.0'
+
+repositories {
+  mavenCentral()
+}
+
+dependencies {
+  implementation project(":offset-manager-api")
+
+  implementation("org.springframework.boot:spring-boot-starter-web")
+  implementation("org.springframework.boot:spring-boot-starter-validation")
+  implementation("org.apache.kafka:kafka-clients:4.1.0")
+  implementation("org.springframework.kafka:spring-kafka:3.3.10")
+
+  compileOnly("org.projectlombok:lombok:1.18.42")
+  annotationProcessor("org.projectlombok:lombok:1.18.42")
+  testCompileOnly("org.projectlombok:lombok:1.18.42")
+  testImplementation platform('org.junit:junit-bom:5.10.0')
+  testImplementation 'org.springframework.boot:spring-boot-starter-test'
+  testImplementation 'org.junit.jupiter:junit-jupiter'
+  testImplementation 'org.assertj:assertj-core:3.27.6'
+  testImplementation 'org.mockito:mockito-core:5.20.0'
+  testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+
+}
+
+test {
+  useJUnitPlatform()
+}
