+++ offset-manager/src/main/java/offsetmanager/OffsetManagerApplication.java
@@ -1,0 +1,8 @@
+package offsetmanager;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OffsetManagerApplication {
+
+}
