+++ offset-manager/src/main/java/offsetmanager/controller/dto/LastOffsetRecordBatchRequest.java
@@ -1,0 +1,11 @@
+package offsetmanager.controller.dto;
+
+import jakarta.validation.constraints.NotEmpty;
+
+import java.util.List;
+
+public record LastOffsetRecordBatchRequest(
+  @NotEmpty
+  List<String> keys
+) {
+}
