+++ offset-manager/src/main/java/offsetmanager/service/dto/LastOffsetRecordResponse.java
@@ -1,0 +1,15 @@
+package offsetmanager.service.dto;
+
+import offsetmanager.domain.OffsetRecord;
+
+public record LastOffsetRecordResponse(
+  String key,
+  long offset
+) {
+  public static LastOffsetRecordResponse from(OffsetRecord offsetRecord) {
+    return new LastOffsetRecordResponse(
+      offsetRecord.key(),
+      offsetRecord.offset()
+    );
+  }
+}
